USB connection and running Etcher as admin
If the USB to a computer, at least in Windows, is not done before the ODROID is powered on and the ums command is executed, Windows does not detect the mass storage.
Also in Windows, if Etcher is run without admin, you get a UAC prompt to approve, and even if approved, Etcher appears to hang. Running Etcher with admin in Windows solves this problem. I spent 3 days trying to flash the ODROID until I figured out the steps :)